Add bug workaround mod+t
This works around a bug where if you start the window manager
in tabbed or stacked mode, the toggle split won't actually work with
a split until a 'layout splitv' or 'layout splith' is issued.
I've just bound it to a spare key so I can fix a restart of the
machine without having to do "i3-msg 'layout splitv'" from a
terminal.